% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_combinations.R
\name{all_combinations}
\alias{all_combinations}
\title{Make all combinations of element vector}
\usage{
all_combinations(elem.bout)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}
}
\value{
data frame with all antecedent and consequent combinations
}
\description{
Make all combinations of element vector
}
\author{
Alex Mielke
}
