% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_of_words.R
\name{bag_of_words}
\alias{bag_of_words}
\title{Checks probabilities of combinations of 2 elements in the dataset based on a 'bag-of-words' approach: user either defines local context (elements within a time gap to the left or right) or select full bout}
\usage{
bag_of_words(elem.bout, elem.time = NULL, gap = c(-1, 1))
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elem.time}{list of vectors of timings of elements. If set to NULL, the order of elements is used with distance of 1}

\item{gap}{vector with two values, one for the start of the bag (compared to element itself) and one with the end (e.g., c(-1,1) starts one second before the element and finishes one second after). Number of elements in this time frame can be variable.}
}
\value{
Function returns a data frame with the element combinations and how often one was in the bag of the other one
}
\description{
Checks probabilities of combinations of 2 elements in the dataset based on a 'bag-of-words' approach: user either defines local context (elements within a time gap to the left or right) or select full bout
}
\author{
Alex Mielke
}
