% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_bag_of_words.R
\name{boot_bag_of_words}
\alias{boot_bag_of_words}
\title{Bootstrap bouts and extract information on expected bag-of-words}
\usage{
boot_bag_of_words(
  elem.bout,
  elem.time,
  measure = c("sum", "prob", "mi"),
  gap = c(-1, 1),
  cores = 2,
  trials = 1000,
  ci.range = c(0.025, 0.975),
  output = "summary"
)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elem.time}{list of vectors of timings of elements. If set to NULL, the order of elements is used with distance of 1}

\item{measure}{which measure to quantify the antecedent - consequent transition probability: 'sum' gives the number of transitions, 'prob' gives conditional probability, 'joint.prob' gives joint probability, 'mi' gives point-wise mutual information}

\item{gap}{number of elements to each side that should be counted in the same 'bag' as the element; if 'all', use whole bout}

\item{cores}{number of cores for parallel randomization (default is 2)}

\item{trials}{number of random trials to calculate bootstrapped values}

\item{ci.range}{range of credible interval extracted from the bootstraps; default is ci.range = c(0.025, 0.975), which means the bootstrapped range excluding the most extreme 2.5 percentage at each tail}

\item{output}{default is 'expected' (give only summary values across all randomizations); can be set to 'raw' (give the value for all randomizations)}
}
\value{
Returns a list with 'sum' (expected sum of transitions), 'prob' (expected probability of transition), 'joint' (expected joint probability), 'p-value' (comparison of observed with all randomizations), 'z' (z value of observed against all randomizations)
}
\description{
Bootstrap bouts and extract information on expected bag-of-words
}
\author{
Alex Mielke
}
