% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_elements.R
\name{boot_elements}
\alias{boot_elements}
\title{Test robustness and distribution of element transitions. Randomly select subsets of the bouts and create bootstrapped information on Antecedent -> consequent transitions. Extracts all different ways (sum of transitions, conditional probability, joint probability, point-wise mutual information), and returns either raw bootstraps or standard deviation, credible range, and coefficient of variance}
\usage{
boot_elements(
  elem.bout,
  antecedent,
  consequent,
  measure = c("sum", "prob", "mi", "joint.prob"),
  it = 20,
  cores = 2,
  trials = 1000,
  ci.range = c(0.025, 0.975),
  output = "summary",
  ran.method = "random"
)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol}

\item{antecedent}{first element in transition}

\item{consequent}{second element in transition}

\item{measure}{which measure to quantify the antecedent - consequent transition probability: 'sum' gives the number of transitions, 'prob' gives conditional probability, 'joint.prob' gives joint probability, 'mi' gives point-wise mutual information}

\item{it}{number of iterations of assignments}

\item{cores}{number of cores for parallel randomization (default is 2)}

\item{trials}{number of random trials to calculate expected values and z, p etc}

\item{ci.range}{range of credible interval extracted from the bootstraps; default is ci.range = c(0.025, 0.975), which means the bootstrapped range excluding the most extreme 2.5 percentage at each tail}

\item{output}{default is 'expected' (give only summary values across all randomizations); can be set to 'raw' (give the value for all randomizations)}

\item{ran.method}{either 'sample' (just sample one of the co-occurring) or 'random' (randomizes their order)}
}
\value{
Returns either raw values from bootstraps or list with credible range, sd, and coefficient of variance for 'sum' (sum of transitions), 'prob' (probability of transition), 'joint.prob' (joint probability), or 'mi' (point-wise mutual information)
}
\description{
Test robustness and distribution of element transitions. Randomly select subsets of the bouts and create bootstrapped information on Antecedent -> consequent transitions. Extracts all different ways (sum of transitions, conditional probability, joint probability, point-wise mutual information), and returns either raw bootstraps or standard deviation, credible range, and coefficient of variance
}
\author{
Alex Mielke
}
