% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_elements.R
\name{change_elements}
\alias{change_elements}
\title{Change element names, allows you to put several elements together}
\usage{
change_elements(elem.bout, elements.to.change, new)
}
\arguments{
\item{elem.bout}{vector of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elements.to.change}{vector of element names to be changed, e.g. c('wave-arm','wiggle-arm','raise-arm')}

\item{new}{new name to give to those elements (e.g., 'move-arm')}
}
\value{
Function returns a list of vectors with renamed elements
}
\description{
Because element names might be overlapping and tied up in co-occurrences, we can't just use str_replace, so everything has to be unlisted, elements renamed, and then put back together
}
\author{
Alex Mielke
}
