% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_elements.R
\name{combine_elements}
\alias{combine_elements}
\title{Take elements that occur together and remove one of them}
\usage{
combine_elements(elem.bout, elements.to.change, new)
}
\arguments{
\item{elem.bout}{vector of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elements.to.change}{vector of element names to be changed, e.g. c('wave-arm','wiggle-arm','raise-arm')}

\item{new}{new name to give to those elements (e.g., 'move-arm')}
}
\value{
Function returns a list of vectors with renamed elements
}
\description{
Take elements that occur together and remove one of them
}
\author{
Alex Mielke
}
