% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_info.R
\name{elem_info}
\alias{elem_info}
\title{Extract information on Antecedent -> consequent transitions. Can be quantified in different ways (sum of transitions, conditional probability, joint probability, point-wise mutual information)}
\usage{
elem_info(
  antecedent,
  consequent,
  elem.bout,
  it,
  measure = c("sum", "prob", "mi", "joint.prob"),
  ran.method = "random"
)
}
\arguments{
\item{antecedent}{first element in transition}

\item{consequent}{second element in transition}

\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{it}{number of iterations of assignments}

\item{measure}{which measure to quantify the antecedent - consequent transition probability: 'sum' gives the number of transitions, 'prob' gives conditional probability, 'joint.prob' gives joint probability, 'mi' gives point-wise mutual information}

\item{ran.method}{either 'sample' (just sample one of the co-occurring) or 'random' (randomizes their order)}
}
\value{
Returns a dataframe with the antecedents, consequents and the information about the transitions
}
\description{
Extract information on Antecedent -> consequent transitions. Can be quantified in different ways (sum of transitions, conditional probability, joint probability, point-wise mutual information)
}
\author{
Alex Mielke
}
