% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_combinations.R
\name{element_combinations}
\alias{element_combinations}
\title{Checks probabilities of combinations of n elements in the dataset (e.g., n = 1: 'slap', n = 2: 'mount:slap', n = 3: 'mount:slap:scream')}
\usage{
element_combinations(elem.bout, lvl = 0, it = 1, ran.method = "random")
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{lvl}{size of combinations of following elements. lvl = 0 (default) means the algorithm finds single elements; lvl = 1 takes dyadic combinations etc}

\item{it}{how many iterations of the random splitting of co-occurring elements should be used? If elements never co-occur, this can be set to 1. If there are many co-occurrences, this number should be high, but this will make things slower}

\item{ran.method}{for co-occurring elements, how should they be split? Can be 'random' (order is randomised), or 'sample' (randomly choose any)}
}
\value{
Function returns a transition matrix. Because it sums over different co-occurrence splits, transitions numbers are not integers
}
\description{
Checks probabilities of combinations of n elements in the dataset (e.g., n = 1: 'slap', n = 2: 'mount:slap', n = 3: 'mount:slap:scream')
}
\examples{
###
}
\author{
Alex Mielke
}
