% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_plot.R
\name{element_plot}
\alias{element_plot}
\title{Create ego-network plot for an element - plots the significant antecedents and consequents for this element}
\usage{
element_plot(
  element,
  antecedent,
  consequent,
  count.antecedent,
  count.consequent,
  observed.probs,
  pvalue,
  cutoff = 5,
  significance = 0.01
)
}
\arguments{
\item{element}{element name}

\item{antecedent}{antecedent element in transition}

\item{consequent}{consequent element in transition}

\item{count.antecedent}{number of times the antecedent occurred (output of 'transition.frame')}

\item{count.consequent}{number of times the consequent occurred (output of 'transition.frame')}

\item{observed.probs}{observed conditional probability to go from antecedent to consequent (output of 'elem.info')}

\item{pvalue}{pvalue of observed against randomized expected probabilities (output of 'randomized.elem.info')}

\item{cutoff}{number of times transitions should have been observed to be included (default is cutoff = 5)}

\item{significance}{level of significance for transitions to be considered significantly more common than expected; default is significance = 0.01}
}
\value{
Returns a ggplot element indicating the antecedents, consequents, and probabilities in each direction
}
\description{
Create ego-network plot for an element - plots the significant antecedents and consequents for this element
}
\examples{
###
}
\author{
Alex Mielke
}
