% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_array.R
\name{mean_array}
\alias{mean_array}
\title{Helper functions that take a list of matrices and sums across them or takes the mean}
\usage{
mean_array(list.matr)
}
\arguments{
\item{list.matr}{list of transition matrices}
}
\value{
mean.array returns a matrix with the mean across all individual matrices. sum.array returns a matrix with the sum across all individual matrices
}
\description{
Helper functions that take a list of matrices and sums across them or takes the mean
}
\examples{
###


## summarise array
}
\author{
Alex Mielke
}
