% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{network_plot}
\alias{network_plot}
\title{Create network plot for element connections - plots the significant antecedents and consequents for all element}
\usage{
network_plot(
  elem.bout,
  elem.time = NULL,
  edge.weight = "transition",
  min.prob = 0.2,
  min.count = 5,
  significance = 0.05,
  hide_unconnected = T,
  link = "unweighted",
  clusters = T,
  plot.bubbles = FALSE,
  ignore.element = NULL,
  title = "network",
  facet = FALSE,
  remove_loops = FALSE,
  gap = c(-2, 2),
  it = 1,
  cores = 5,
  plot.layout = "fr",
  ran.method = "random"
)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elem.time}{list of vectors of timings of elements. If set to NULL, the order of elements is used with distance of 1}

\item{edge.weight}{either 'transition', in which case the edges are defined by the probability of A transitioning to B, or 'bag', in which case a 'bag-of-words' approach is chosen}

\item{min.prob}{numeric value, suggesting the probability at which a combination should at least occur to be displayed#'}

\item{min.count}{numeric value, suggesting how many times a combination should at least occur to be displayed}

\item{significance}{level of significance for transitions to be considered significantly more common than expected; default is sig = 0.01}

\item{hide_unconnected}{if TRUE, then the nodes that do not have any significant connections will be hidden in the plot}

\item{link}{determines how nodes/elements are connected. 'unweighted' gives a 1 to significant connections and 0 to all others; 'weighted' gives the difference between observed and expected probability of co-occurrence; 'raw' just uses the observed probability of co-occurrence; 'SRI' uses the simple ratio index/affinity (probability of co-occurrence/ (probabilities of each element and the combination))}

\item{clusters}{should clusters be calculated and added?}

\item{plot.bubbles}{if TRUE, then the nodes in the network plots will be surrounded by bubbles; if FALSE, the edges connect the names directly}

\item{ignore.element}{vector of elements that will not be considered for the network, e.g. because they are too common or too rare or their interpretation is not relevant here}

\item{title}{string, will be the title of the resulting plot}

\item{facet}{if TRUE, plot will be split by clusters; default is FALSE}

\item{remove_loops}{if FALSE, significant connections of elements with themselves are shown; default is FALSE}

\item{gap}{number of elements to each side that should be counted in the same 'bag' as the element; if 'all', use whole bout}

\item{it}{how many iterations of the random splitting of co-occurring elements should be used? If elements never co-occur, this can be set to 1. If there are many co-occurrences, this number should be high, but this will make things slower}

\item{cores}{number of cores for parallel execution}

\item{plot.layout}{one of the igraph layout options; works okay with 'fr', 'lgl', 'dh', 'nicely', 'kk', or 'graphopt', depending on the number of elements}

\item{ran.method}{for co-occurring elements, how should they be split? Can be 'random' (order is randomised), or 'sample' (randomly choose any)}
}
\value{
Returns a ggplot element indicating the antecedents, descendents, and probabilities in each direction; also returns an igraph graph and the centrality measures for each element
}
\description{
Create network plot for element connections - plots the significant antecedents and consequents for all element
}
\examples{
###
}
\author{
Alex Mielke
}
