% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_bouts.R
\name{randomize_bouts}
\alias{randomize_bouts}
\title{Take sequence within bout and randomize sequence, keeping the number of elements the same and either controlling for the overall probability across bouts, or control the elements within bouts}
\usage{
randomize_bouts(elem.bout, type = "across", unlisting = T)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol. Function returns same vector, but with co-occurring elements split.}

\item{type}{how to determine probability of elements for randomization. If 'across', frequencies of elements are the same across all bouts, while 'within' shuffles elements within a bout}

\item{unlisting}{should output be unlisted or retain original list structure?}
}
\value{
Function returns a list of vectors with the randomized, split elements
}
\description{
The \code{\link{randomize_bouts}} takes the lists of bouts and randomizes the elements while keeping the number of elements per bout
}
\author{
Alex Mielke
}
