% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_bag_of_words.R
\name{randomized_bag_of_words}
\alias{randomized_bag_of_words}
\title{Randomize bouts and extract information on expected bag-of-words}
\usage{
randomized_bag_of_words(
  elem.bout,
  elem.time,
  observed = NULL,
  gap = c(-1, 1),
  cores = 2,
  trials = 1000,
  output = "expected"
)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elem.time}{list of vectors of timings of elements. If set to NULL, the order of elements is used with distance of 1}

\item{observed}{vector of observed probabilities to calculate z and p-values}

\item{gap}{number of elements to each side that should be counted in the same 'bag' as the element; if 'all', use whole bout}

\item{cores}{number of cores for parallel randomization (default is 2)}

\item{trials}{number of random trials to calculate expected values and z, p etc}

\item{output}{default is 'expected' (give only summary values across all randomizations); can be set to 'raw' (give the value for all randomizations)}
}
\value{
Returns a list with 'sum' (expected sum of transitions), 'prob' (expected probability of transition), 'joint' (expected joint probability), 'p-value' (comparison of observed with all randomizations), 'z' (z value of observed against all randomizations)
}
\description{
Randomize bouts and extract information on expected bag-of-words
}
\examples{
###
}
\author{
Alex Mielke
}
