% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_elem_info.R
\name{randomized_elem_info}
\alias{randomized_elem_info}
\title{Randomize bouts and extract information on expected Antecedent -> consequent transitions. Extracts all different ways (sum of transitions, conditional probability, joint probability, point-wise mutual information)}
\usage{
randomized_elem_info(
  elem.bout,
  antecedent,
  consequent,
  observed = NULL,
  it = 20,
  cores = 2,
  trials = 1000,
  type = "across",
  output = "expected",
  ran.method = "random"
)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{antecedent}{first element in transition}

\item{consequent}{second element in transition}

\item{observed}{vector of observed probabilities to calculate z and p-values}

\item{it}{number of iterations of assignments}

\item{cores}{number of cores for parallel randomization (default is 2)}

\item{trials}{number of random trials to calculate expected values and z, p etc}

\item{type}{how to determine probability of elements for randomization. If 'across', frequencies of elements are the same across all bouts, while 'within' shuffles elements within a bout}

\item{output}{default is 'expected' (give only summary values across all randomizations); can be set to 'raw' (give the value for all randomizations)}

\item{ran.method}{for co-occurring elements, how should they be split? Can be 'random' (order is randomised), or 'sample' (randomly choose any)}
}
\value{
Returns a list with 'sum' (expected sum of transitions), 'prob' (expected probability of transition), 'joint' (expected joint probability), 'p-value' (comparison of observed with all randomizations), 'z' (z value of observed against all randomizations)
}
\description{
Randomize bouts and extract information on expected Antecedent -> consequent transitions. Extracts all different ways (sum of transitions, conditional probability, joint probability, point-wise mutual information)
}
\examples{
###
}
\author{
Alex Mielke
}
