% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions_frame.R
\name{transitions_frame}
\alias{transitions_frame}
\title{Create dataframe that summarises how often each individual element occurs in the dataset; basis for many of the other functions}
\usage{
transitions_frame(elem.bout, elements = NULL, all.possible = FALSE)
}
\arguments{
\item{elem.bout}{list of vectors of elements. Co-occurring elements are marked by the percentage symbol.}

\item{elements}{vector of all elements that should be included in the analysis; if 'NULL', the function takes all elements that occur in the elem.bout list}

\item{all.possible}{should output contain all possible combinations of elements, or only those that occurred at least 1 time?}
}
\value{
Function returns a data frame with each antecedent, consequent, and their individual occurrence stats
}
\description{
Create dataframe that summarises how often each individual element occurs in the dataset; basis for many of the other functions
}
\examples{
###
}
\author{
Alex Mielke
}
