% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_list.R
\name{unlist_list}
\alias{unlist_list}
\title{Take list of sequences within bout and randomly assign sequence for co-occurring elements (marked by percentage symbol)}
\usage{
unlist_list(elem.bout, method = "sample")
}
\arguments{
\item{elem.bout}{list of bouts of elements. Co-occurring elements are marked by the percentage sign. Function returns same list of vectors, but with co-occurring elements split.}

\item{method}{either 'sample' (just sample one of the co-occurring); 'random' (randomizes their order); or 'optimal' (calculates probabilities outside of this case and assigns based on those)

The resulting object is the basis for most other functions in this package.}
}
\value{
Function returns a list of vectors with the split elements
}
\description{
Takes a list of the vector consisting of bouts of elements in their order of occurrence and unlists elements that occur at the same time, randomly.
}
\author{
Alex Mielke
}
