% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipfs_plot.R
\name{zipf_plot}
\alias{zipf_plot}
\title{Plots expected Zipf's distribution against the actual occurrence of elements in the dataset}
\usage{
zipf_plot(element, occurrance, zipf = T, title = "Zipfs Law Visualization")
}
\arguments{
\item{element}{vector with element names}

\item{occurrance}{vector with sums of occurrance for each element}
}
\value{
Function returns the entropy value in bits for each of the levels
}
\description{
Expected Zipf's distribution is calculated as (Frequency of most common)/Rank
}
\examples{
###
}
\author{
Alex Mielke
}
